refactor: Divide handler and logic functions
